requirements.txt
ctransformers[cuda]
transformers
torch
discord.py
python-dotenv
gdown
flask
